Remove d2l-my-courses-behavior, add functionality back into d2l-my-courses-container
